requirements.txt
openai-whisper
torch
pyttsx3
sounddevice
scipy
Pillow
python-dotenv